refactor(진성진): 폴더 분리 (constants, utils)
