+++ signin.html
@@ -74,6 +74,5 @@
       </div>
     </main>
   </body>
-  <script type="module" src="src/js/validator.mjs"></script>
   <script type="module" src="src/js/signin.mjs"></script>
 </html>
