+++ signup.html
@@ -108,6 +108,5 @@
       </div>
     </main>
   </body>
-  <script type="module" src="src/js/validator.mjs"></script>
   <script type="module" src="src/js/signup.mjs"></script>
 </html>
